Refactor formatDuration utility to a shared module
- Moved `formatDuration` from `ReaderPage.kt` to `core:base` as a shared utility.
- Added unit tests for `formatDuration` in `core:base`.
- Updated `ReaderPage.kt` to use the shared utility.

This improves code reusability and maintainability by centralizing the duration formatting logic.
