+++ core/base/src/commonMain/kotlin/dev/sasikanth/rss/reader/util/DurationExt.kt
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2026 Sasikanth Miriyampalli
+ *
+ * Licensed under the GPL, Version 3.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.gnu.org/licenses/gpl-3.0.en.html
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package dev.sasikanth.rss.reader.util
+
+fun formatDuration(duration: Long): String {
+  val seconds = (duration / 1000) % 60
+  val minutes = (duration / (1000 * 60)) % 60
+  val hours = (duration / (1000 * 60 * 60))
+
+  return if (hours > 0) {
+    "${hours.toString().padStart(2, '0')}:${minutes.toString().padStart(2, '0')}:${seconds.toString().padStart(2, '0')}"
+  } else {
+    "${minutes.toString().padStart(2, '0')}:${seconds.toString().padStart(2, '0')}"
+  }
+}
